% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{calculate.RF}
\alias{calculate.RF}
\alias{calculate.RF,RFclust.SGE-method}
\title{description of function calculate.RF}
\usage{
calculate.RF(dat, datRF = NULL, mtry1 = 3, no.rep = 20, no.tree = 500,
  addcl1 = TRUE, addcl2 = FALSE, imp = T, oob.prox1 = T, max.syn = 50)
}
\arguments{
\item{datRF}{the summary object is a list of random forest return values default= NULL}

\item{mtry1}{Number of variables randomly sampled as candidates at each split. default=3}

\item{no.rep}{how many times shoul the random forest function be called default= 20}

\item{no.tree}{the numer of trees in the random forest function default= 500}

\item{addcl1}{synthetic data created by the synthetic1 function default=TRUE}

\item{addcl2}{synthetic data created by the synthetic2 function default=FALSE}

\item{imp}{whether to calculate the importance default=T}

\item{oob.prox1}{leave it like that default=T}

\item{max.syn}{max synthetic data width default=50}
}
\description{
This function is the threadable random forest function.

The return of this fucntion should be stored to a file using save.RF

and later on during the clean up of the different threads be read using read.RF

The return list can be processed using the RFdist method
}

