% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{collect.garbage}
\alias{collect.garbage}
\alias{collect.garbage,RFclust.SGE-method}
\title{description of function collect.garbage}
\usage{
collect.garbage(x)
}
\description{
The following function collects garbage until the memory is clean.

Usage: 1. immediately call this function after you call a function or

2. rm()
}
