% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{locked}
\alias{locked}
\alias{locked,RFclust.SGE-method}
\title{description of function locked}
\usage{
locked(filename)
}
\arguments{
\item{filename}{lock this file}
}
\description{
simple check for a file lock (threading)
}
