% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{pamNew}
\alias{pamNew}
\alias{pamNew,RFclust.SGE-method}
\title{description of function pamNew}
\usage{
pamNew(x, k, diss1 = inherits(x, "dist"), metric1 = "euclidean")
}
\arguments{
\item{x}{the distRF object obtained from RFdist()}

\item{k}{number of exprected clusters}

\item{diss1}{TEXT MISSING default= inherits(x, "dist")}

\item{metric1}{TEXT MISSING default= "euclidean"}
}
\description{
A new pam clustering function which corrects the clustering membership based on the sillhouette strength.

The clustering membership of an observation with a negative sillhouette strength is reassigned to its

neighboring cluster.

The inputs of the function are similar to the original 'pam' function.

The function returns a vector of clustering labels.

Copyright 2003 Tao Shi and Steve Horvath (last modified 10/31/03)
}
