% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{Rand}
\alias{Rand}
\alias{Rand,RFclust.SGE-method}
\title{description of function Rand}
\usage{
Rand(tab, adjust = T)
}
\arguments{
\item{tab}{the data table}

\item{adjust}{just leave it ... default=T}
}
\description{
The function computes the (adjusted) Rand index between two partitions

Copyright Steve Horvath and Luohua Jiang, UCLA, 2003
}
