% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{release.lock}
\alias{release.lock}
\alias{release.lock,RFclust.SGE-method}
\title{description of function release_lock}
\usage{
release.lock(filename)
}
\arguments{
\item{filename}{the locked file}
}
\description{
releases the lock of a file (threading)
}

