% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{RFdist}
\alias{RFdist}
\alias{RFdist,RFclust.SGE-method}
\title{description of function RFdist}
\usage{
RFdist(Rf.data, datRF, imp = T, no.tree, proxConver = F)
}
\arguments{
\item{Rf.data}{TEXT MISSING}

\item{datRF}{TEXT MISSING}

\item{imp}{TEXT MISSING default=T}

\item{no.tree}{TEXT MISSING}

\item{proxConver}{TEXT MISSING default=F}
}
\description{
Initially this function did perform the unsupervised clustering,

but the most calculation has been exported to the threadable function

\link{calculate.RF}
}
