% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{save.RF}
\alias{save.RF}
\alias{save.RF,RFclust.SGE-method}
\title{description of function save.RF}
\usage{
save.RF(Rf.data, fname)
}
\arguments{
\item{Rf.data}{the return value from a calculate.RF call}

\item{fname}{the file to store the data in}
}
\description{
Store a calculate.RF result in a file for later usage (threading)
}
