% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFclust.SGE.R
\docType{methods}
\name{set.lock}
\alias{set.lock}
\alias{set.lock,RFclust.SGE-method}
\title{description of function set_lock}
\usage{
set.lock(filename)
}
\arguments{
\item{filename}{the locked file}
}
\description{
set a lock for a file (threading)
}

