% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,RFclust.SGE-method}
\title{description of function show}
\usage{
\S4method{show}{RFclust.SGE}(object)
}
\arguments{
\item{object}{the RFclust.SGE object}
}
\description{
print the object
}

