% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{writeRscript}
\alias{writeRscript}
\alias{writeRscript,RFclust.SGE-method}
\title{description of function writeRscript}
\usage{
writeRscript(
  x,
  filename,
  ntree = 500,
  nforest = 500,
  run = TRUE,
  srcObj,
  total.n = NULL
)
}
\arguments{
\item{filename}{the filename to save the R script to (has to be unique for the analysis!)}

\item{ntree}{the number of trees to grow}

\item{nforest}{the nuber of forests to create}

\item{total.n}{the total number of forests to calcualte}
}
\value{
a filename for the expected data
}
\description{
run the random forest calculations returning the density matrix

at the moment without SGE support and single core
}
