% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{writeSGEscript}
\alias{writeSGEscript}
\alias{writeSGEscript,RFclust.SGE-method}
\title{description of function writeSGEscript}
\usage{
writeSGEscript(x, ntree = 500, nforest = 500)
}
\arguments{
\item{x}{the RFclust.SGE object}

\item{ntree}{the number of trees to grow}

\item{nforest}{the nuber of forests to create}
}
\value{
a distRF object to be analyzed by pamNew
}
\description{
run the random forest calculations returning the density matrix

at the moment without SGE support and single core
}

