% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{RFclust.SGE}
\alias{RFclust.SGE}
\alias{RFclust.SGE,RFclust.SGE-method}
\title{description of function RFclust.SGE}
\usage{
RFclust.SGE(dat, tmp.path = "", email = "", slices = 32, SGE = FALSE)
}
\arguments{
\item{dat}{data frame or matrix containing all expression data}

\item{tmp.path}{where to store the temporaray files}

\item{slices}{how many threads to span}

\item{SGE}{whether to use the Sun Grid Engine to calcualate}
}
\value{
A new RFclust.SGE object
}
\description{
create a new RFclust.SGE object. The clustering will be performed on the columns of the data.frame.
}

