% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{package}
\name{RFclust.SGE}
\alias{RFclust.SGE}
\alias{RFclust.SGE-package}
\title{RFclust.SGE}
\source{
\url{https://labs.genetics.ucla.edu/horvath/RFclustering/RFclustering.htm}
}
\description{
An S4 class to run unsupervides clustering based on random forrest predictors using the Sun Grid Engine.
}
\section{Slots}{

\describe{
\item{\code{dat}}{the data that should be clustered (data.frame with column ==  samples rows ==  observations)}

\item{\code{email}}{the users email for the SGE report}

\item{\code{slices}}{the number of slices the data should be analyzed in}

\item{\code{tmp.path}}{the temp path for the clustering results}

\item{\code{SGE}}{run using Sun Grid Engine (default= F)}

\item{\code{distRF}}{the density data (internal use only)}

\item{\code{RFfiles}}{an internal list of RF files (internal use only)}

\item{\code{name}}{the name for this object that will be used to identify the data object in the spawned processes}
}}

