% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{writeSGEscript}
\alias{writeSGEscript}
\alias{writeSGEscript,RFclust.SGE-method}
\title{description of function writeSGEscript}
\usage{
writeSGEscript(x, filename, cmd)
}
\arguments{
\item{x}{the RFclust.SGE object}

\item{filename}{the base filename for the script (path and .sh will be added!)}

\item{cmd}{the command to include in the SGE script. Make sure, that all path entries are valid on the nodes!}
}
\value{
a distRF object to be analyzed by pamNew
}
\description{
run the random forest calculations returning the density matrix

at the moment without SGE support and single core
}

